MAINT: Split generic.py
