+++ PyPDF2/generic/_data_structures.py
@@ -26,163 +26,52 @@
 # POSSIBILITY OF SUCH DAMAGE.
 
 
-"""Implementation of generic PDF objects (dictionary, number, string, ...)."""
 __author__ = "Mathieu Fenniak"
 __author_email__ = "biziqe@mathieu.fenniak.net"
 
-import codecs
 import decimal
-import hashlib
 import logging
 import re
 from enum import IntFlag
 from io import BytesIO
-from typing import (
-    Any,
-    Callable,
-    Dict,
-    Iterable,
-    List,
-    Optional,
-    Tuple,
-    Union,
-    cast,
-)
+from typing import Any, Dict, Iterable, List, Optional, Tuple, Union, cast
 
-from ._codecs import (  # noqa: rev_encoding
-    _pdfdoc_encoding,
-    _pdfdoc_encoding_rev,
-    rev_encoding,
-)
-from ._utils import (
+from .._utils import (
     WHITESPACES,
     StreamType,
     b_,
     deprecate_no_replacement,
     deprecate_with_replacement,
     hex_str,
-    hexencode,
     logger_warning,
     read_non_whitespace,
     read_until_regex,
     skip_over_comment,
-    str_,
 )
-from .constants import CheckboxRadioButtonAttributes, FieldDictionaryAttributes
-from .constants import FilterTypes as FT
-from .constants import StreamAttributes as SA
-from .constants import TypArguments as TA
-from .constants import TypFitArguments as TF
-from .errors import STREAM_TRUNCATED_PREMATURELY, PdfReadError, PdfStreamError
+from ..constants import (
+    CheckboxRadioButtonAttributes,
+    FieldDictionaryAttributes,
+)
+from ..constants import FilterTypes as FT
+from ..constants import StreamAttributes as SA
+from ..constants import TypArguments as TA
+from ..constants import TypFitArguments as TF
+from ..errors import STREAM_TRUNCATED_PREMATURELY, PdfReadError, PdfStreamError
+from ._base import (
+    BooleanObject,
+    FloatObject,
+    IndirectObject,
+    NameObject,
+    NullObject,
+    NumberObject,
+    PdfObject,
+)
+from ._utils import read_hex_string_from_stream, read_string_from_stream
 
 logger = logging.getLogger(__name__)
 ObjectPrefix = b"/<[tf(n%"
 NumberSigns = b"+-"
 IndirectPattern = re.compile(rb"[+-]?(\d+)\s+(\d+)\s+R[^a-zA-Z]")
-
-
-class PdfObject:
-    # function for calculating a hash value
-    hash_func: Callable[..., "hashlib._Hash"] = hashlib.sha1
-
-    def hash_value_data(self) -> bytes:
-        return ("%s" % self).encode()
-
-    def hash_value(self) -> bytes:
-        return (
-            "%s:%s"
-            % (
-                self.__class__.__name__,
-                self.hash_func(self.hash_value_data()).hexdigest(),
-            )
-        ).encode()
-
-    def get_object(self) -> Optional["PdfObject"]:
-        """Resolve indirect references."""
-        return self
-
-    def getObject(self) -> Optional["PdfObject"]:  # pragma: no cover
-        deprecate_with_replacement("getObject", "get_object")
-        return self.get_object()
-
-    def write_to_stream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:
-        raise NotImplementedError
-
-
-class NullObject(PdfObject):
-    def write_to_stream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:
-        stream.write(b"null")
-
-    @staticmethod
-    def read_from_stream(stream: StreamType) -> "NullObject":
-        nulltxt = stream.read(4)
-        if nulltxt != b"null":
-            raise PdfReadError("Could not read Null object")
-        return NullObject()
-
-    def writeToStream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:  # pragma: no cover
-        deprecate_with_replacement("writeToStream", "write_to_stream")
-        self.write_to_stream(stream, encryption_key)
-
-    def __repr__(self) -> str:
-        return "NullObject"
-
-    @staticmethod
-    def readFromStream(stream: StreamType) -> "NullObject":  # pragma: no cover
-        deprecate_with_replacement("readFromStream", "read_from_stream")
-        return NullObject.read_from_stream(stream)
-
-
-class BooleanObject(PdfObject):
-    def __init__(self, value: Any) -> None:
-        self.value = value
-
-    def __eq__(self, __o: object) -> bool:
-        if isinstance(__o, BooleanObject):
-            return self.value == __o.value
-        elif isinstance(__o, bool):
-            return self.value == __o
-        else:
-            return False
-
-    def __repr__(self) -> str:
-        return "True" if self.value else "False"
-
-    def write_to_stream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:
-        if self.value:
-            stream.write(b"true")
-        else:
-            stream.write(b"false")
-
-    def writeToStream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:  # pragma: no cover
-        deprecate_with_replacement("writeToStream", "write_to_stream")
-        self.write_to_stream(stream, encryption_key)
-
-    @staticmethod
-    def read_from_stream(stream: StreamType) -> "BooleanObject":
-        word = stream.read(4)
-        if word == b"true":
-            return BooleanObject(True)
-        elif word == b"fals":
-            stream.read(1)
-            return BooleanObject(False)
-        else:
-            raise PdfReadError("Could not read Boolean object")
-
-    @staticmethod
-    def readFromStream(stream: StreamType) -> "BooleanObject":  # pragma: no cover
-        deprecate_with_replacement("readFromStream", "read_from_stream")
-        return BooleanObject.read_from_stream(stream)
 
 
 class ArrayObject(list, PdfObject):
@@ -241,423 +130,6 @@ class ArrayObject(list, PdfObject):
         return ArrayObject.read_from_stream(stream, pdf)
 
 
-class IndirectObject(PdfObject):
-    def __init__(self, idnum: int, generation: int, pdf: Any) -> None:  # PdfReader
-        self.idnum = idnum
-        self.generation = generation
-        self.pdf = pdf
-
-    def get_object(self) -> Optional[PdfObject]:
-        obj = self.pdf.get_object(self)
-        if obj is None:
-            return None
-        return obj.get_object()
-
-    def __repr__(self) -> str:
-        return f"IndirectObject({self.idnum!r}, {self.generation!r}, {id(self.pdf)})"
-
-    def __eq__(self, other: Any) -> bool:
-        return (
-            other is not None
-            and isinstance(other, IndirectObject)
-            and self.idnum == other.idnum
-            and self.generation == other.generation
-            and self.pdf is other.pdf
-        )
-
-    def __ne__(self, other: Any) -> bool:
-        return not self.__eq__(other)
-
-    def write_to_stream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:
-        stream.write(b_(f"{self.idnum} {self.generation} R"))
-
-    def writeToStream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:  # pragma: no cover
-        deprecate_with_replacement("writeToStream", "write_to_stream")
-        self.write_to_stream(stream, encryption_key)
-
-    @staticmethod
-    def read_from_stream(stream: StreamType, pdf: Any) -> "IndirectObject":  # PdfReader
-        idnum = b""
-        while True:
-            tok = stream.read(1)
-            if not tok:
-                raise PdfStreamError(STREAM_TRUNCATED_PREMATURELY)
-            if tok.isspace():
-                break
-            idnum += tok
-        generation = b""
-        while True:
-            tok = stream.read(1)
-            if not tok:
-                raise PdfStreamError(STREAM_TRUNCATED_PREMATURELY)
-            if tok.isspace():
-                if not generation:
-                    continue
-                break
-            generation += tok
-        r = read_non_whitespace(stream)
-        if r != b"R":
-            raise PdfReadError(
-                f"Error reading indirect object reference at byte {hex_str(stream.tell())}"
-            )
-        return IndirectObject(int(idnum), int(generation), pdf)
-
-    @staticmethod
-    def readFromStream(
-        stream: StreamType, pdf: Any  # PdfReader
-    ) -> "IndirectObject":  # pragma: no cover
-        deprecate_with_replacement("readFromStream", "read_from_stream")
-        return IndirectObject.read_from_stream(stream, pdf)
-
-
-class FloatObject(decimal.Decimal, PdfObject):
-    def __new__(
-        cls, value: Union[str, Any] = "0", context: Optional[Any] = None
-    ) -> "FloatObject":
-        try:
-            return decimal.Decimal.__new__(cls, str_(value), context)
-        except Exception:
-            try:
-                return decimal.Decimal.__new__(cls, str(value))
-            except decimal.InvalidOperation:
-                # If this isn't a valid decimal (happens in malformed PDFs)
-                # fallback to 0
-                logger_warning(f"Invalid FloatObject {value}", __name__)
-                return decimal.Decimal.__new__(cls, "0")
-
-    def __repr__(self) -> str:
-        if self == self.to_integral():
-            return str(self.quantize(decimal.Decimal(1)))
-        else:
-            # Standard formatting adds useless extraneous zeros.
-            o = f"{self:.5f}"
-            # Remove the zeros.
-            while o and o[-1] == "0":
-                o = o[:-1]
-            return o
-
-    def as_numeric(self) -> float:
-        return float(repr(self).encode("utf8"))
-
-    def write_to_stream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:
-        stream.write(repr(self).encode("utf8"))
-
-    def writeToStream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:  # pragma: no cover
-        deprecate_with_replacement("writeToStream", "write_to_stream")
-        self.write_to_stream(stream, encryption_key)
-
-
-class NumberObject(int, PdfObject):
-    NumberPattern = re.compile(b"[^+-.0-9]")
-
-    def __new__(cls, value: Any) -> "NumberObject":
-        val = int(value)
-        try:
-            return int.__new__(cls, val)
-        except OverflowError:
-            return int.__new__(cls, 0)
-
-    def as_numeric(self) -> int:
-        return int(repr(self).encode("utf8"))
-
-    def write_to_stream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:
-        stream.write(repr(self).encode("utf8"))
-
-    def writeToStream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:  # pragma: no cover
-        deprecate_with_replacement("writeToStream", "write_to_stream")
-        self.write_to_stream(stream, encryption_key)
-
-    @staticmethod
-    def read_from_stream(stream: StreamType) -> Union["NumberObject", FloatObject]:
-        num = read_until_regex(stream, NumberObject.NumberPattern)
-        if num.find(b".") != -1:
-            return FloatObject(num)
-        return NumberObject(num)
-
-    @staticmethod
-    def readFromStream(
-        stream: StreamType,
-    ) -> Union["NumberObject", FloatObject]:  # pragma: no cover
-        deprecate_with_replacement("readFromStream", "read_from_stream")
-        return NumberObject.read_from_stream(stream)
-
-
-def readHexStringFromStream(
-    stream: StreamType,
-) -> Union["TextStringObject", "ByteStringObject"]:  # pragma: no cover
-    deprecate_with_replacement(
-        "readHexStringFromStream", "read_hex_string_from_stream", "4.0.0"
-    )
-    return read_hex_string_from_stream(stream)
-
-
-def read_hex_string_from_stream(
-    stream: StreamType,
-    forced_encoding: Union[None, str, List[str], Dict[int, str]] = None,
-) -> Union["TextStringObject", "ByteStringObject"]:
-    stream.read(1)
-    txt = ""
-    x = b""
-    while True:
-        tok = read_non_whitespace(stream)
-        if not tok:
-            raise PdfStreamError(STREAM_TRUNCATED_PREMATURELY)
-        if tok == b">":
-            break
-        x += tok
-        if len(x) == 2:
-            txt += chr(int(x, base=16))
-            x = b""
-    if len(x) == 1:
-        x += b"0"
-    if len(x) == 2:
-        txt += chr(int(x, base=16))
-    return create_string_object(b_(txt), forced_encoding)
-
-
-def readStringFromStream(
-    stream: StreamType,
-    forced_encoding: Union[None, str, List[str], Dict[int, str]] = None,
-) -> Union["TextStringObject", "ByteStringObject"]:  # pragma: no cover
-    deprecate_with_replacement(
-        "readStringFromStream", "read_string_from_stream", "4.0.0"
-    )
-    return read_string_from_stream(stream, forced_encoding)
-
-
-def read_string_from_stream(
-    stream: StreamType,
-    forced_encoding: Union[None, str, List[str], Dict[int, str]] = None,
-) -> Union["TextStringObject", "ByteStringObject"]:
-    tok = stream.read(1)
-    parens = 1
-    txt = b""
-    while True:
-        tok = stream.read(1)
-        if not tok:
-            raise PdfStreamError(STREAM_TRUNCATED_PREMATURELY)
-        if tok == b"(":
-            parens += 1
-        elif tok == b")":
-            parens -= 1
-            if parens == 0:
-                break
-        elif tok == b"\\":
-            tok = stream.read(1)
-            escape_dict = {
-                b"n": b"\n",
-                b"r": b"\r",
-                b"t": b"\t",
-                b"b": b"\b",
-                b"f": b"\f",
-                b"c": rb"\c",
-                b"(": b"(",
-                b")": b")",
-                b"/": b"/",
-                b"\\": b"\\",
-                b" ": b" ",
-                b"%": b"%",
-                b"<": b"<",
-                b">": b">",
-                b"[": b"[",
-                b"]": b"]",
-                b"#": b"#",
-                b"_": b"_",
-                b"&": b"&",
-                b"$": b"$",
-            }
-            try:
-                tok = escape_dict[tok]
-            except KeyError:
-                if tok.isdigit():
-                    # "The number ddd may consist of one, two, or three
-                    # octal digits; high-order overflow shall be ignored.
-                    # Three octal digits shall be used, with leading zeros
-                    # as needed, if the next character of the string is also
-                    # a digit." (PDF reference 7.3.4.2, p 16)
-                    for _ in range(2):
-                        ntok = stream.read(1)
-                        if ntok.isdigit():
-                            tok += ntok
-                        else:
-                            stream.seek(-1, 1)  # ntok has to be analysed
-                            break
-                    tok = b_(chr(int(tok, base=8)))
-                elif tok in b"\n\r":
-                    # This case is  hit when a backslash followed by a line
-                    # break occurs.  If it's a multi-char EOL, consume the
-                    # second character:
-                    tok = stream.read(1)
-                    if tok not in b"\n\r":
-                        stream.seek(-1, 1)
-                    # Then don't add anything to the actual string, since this
-                    # line break was escaped:
-                    tok = b""
-                else:
-                    msg = rf"Unexpected escaped string: {tok.decode('utf8')}"
-                    logger_warning(msg, __name__)
-        txt += tok
-    return create_string_object(txt, forced_encoding)
-
-
-class ByteStringObject(bytes, PdfObject):
-    """
-    Represents a string object where the text encoding could not be determined.
-    This occurs quite often, as the PDF spec doesn't provide an alternate way to
-    represent strings -- for example, the encryption data stored in files (like
-    /O) is clearly not text, but is still stored in a "String" object.
-    """
-
-    @property
-    def original_bytes(self) -> bytes:
-        """For compatibility with TextStringObject.original_bytes."""
-        return self
-
-    def write_to_stream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:
-        bytearr = self
-        if encryption_key:
-            from ._security import RC4_encrypt
-
-            bytearr = RC4_encrypt(encryption_key, bytearr)  # type: ignore
-        stream.write(b"<")
-        stream.write(hexencode(bytearr))
-        stream.write(b">")
-
-    def writeToStream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:  # pragma: no cover
-        deprecate_with_replacement("writeToStream", "write_to_stream")
-        self.write_to_stream(stream, encryption_key)
-
-
-class TextStringObject(str, PdfObject):
-    """
-    Represents a string object that has been decoded into a real unicode string.
-    If read from a PDF document, this string appeared to match the
-    PDFDocEncoding, or contained a UTF-16BE BOM mark to cause UTF-16 decoding to
-    occur.
-    """
-
-    autodetect_pdfdocencoding = False
-    autodetect_utf16 = False
-
-    @property
-    def original_bytes(self) -> bytes:
-        """
-        It is occasionally possible that a text string object gets created where
-        a byte string object was expected due to the autodetection mechanism --
-        if that occurs, this "original_bytes" property can be used to
-        back-calculate what the original encoded bytes were.
-        """
-        return self.get_original_bytes()
-
-    def get_original_bytes(self) -> bytes:
-        # We're a text string object, but the library is trying to get our raw
-        # bytes.  This can happen if we auto-detected this string as text, but
-        # we were wrong.  It's pretty common.  Return the original bytes that
-        # would have been used to create this object, based upon the autodetect
-        # method.
-        if self.autodetect_utf16:
-            return codecs.BOM_UTF16_BE + self.encode("utf-16be")
-        elif self.autodetect_pdfdocencoding:
-            return encode_pdfdocencoding(self)
-        else:
-            raise Exception("no information about original bytes")
-
-    def write_to_stream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:
-        # Try to write the string out as a PDFDocEncoding encoded string.  It's
-        # nicer to look at in the PDF file.  Sadly, we take a performance hit
-        # here for trying...
-        try:
-            bytearr = encode_pdfdocencoding(self)
-        except UnicodeEncodeError:
-            bytearr = codecs.BOM_UTF16_BE + self.encode("utf-16be")
-        if encryption_key:
-            from ._security import RC4_encrypt
-
-            bytearr = RC4_encrypt(encryption_key, bytearr)
-            obj = ByteStringObject(bytearr)
-            obj.write_to_stream(stream, None)
-        else:
-            stream.write(b"(")
-            for c in bytearr:
-                if not chr(c).isalnum() and c != b" ":
-                    # This:
-                    #   stream.write(b_(rf"\{c:0>3o}"))
-                    # gives
-                    #   https://github.com/davidhalter/parso/issues/207
-                    stream.write(b_("\\%03o" % c))
-                else:
-                    stream.write(b_(chr(c)))
-            stream.write(b")")
-
-    def writeToStream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:  # pragma: no cover
-        deprecate_with_replacement("writeToStream", "write_to_stream")
-        self.write_to_stream(stream, encryption_key)
-
-
-class NameObject(str, PdfObject):
-    delimiter_pattern = re.compile(rb"\s+|[\(\)<>\[\]{}/%]")
-    surfix = b"/"
-
-    def write_to_stream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:
-        stream.write(b_(self))
-
-    def writeToStream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:  # pragma: no cover
-        deprecate_with_replacement("writeToStream", "write_to_stream")
-        self.write_to_stream(stream, encryption_key)
-
-    @staticmethod
-    def read_from_stream(stream: StreamType, pdf: Any) -> "NameObject":  # PdfReader
-        name = stream.read(1)
-        if name != NameObject.surfix:
-            raise PdfReadError("name read error")
-        name += read_until_regex(stream, NameObject.delimiter_pattern, ignore_eof=True)
-        try:
-            try:
-                ret = name.decode("utf-8")
-            except (UnicodeEncodeError, UnicodeDecodeError):
-                ret = name.decode("gbk")
-            return NameObject(ret)
-        except (UnicodeEncodeError, UnicodeDecodeError) as e:
-            # Name objects should represent irregular characters
-            # with a '#' followed by the symbol's hex number
-            if not pdf.strict:
-                logger_warning("Illegal character in Name Object", __name__)
-                return NameObject(name)
-            else:
-                raise PdfReadError("Illegal character in Name Object") from e
-
-    @staticmethod
-    def readFromStream(
-        stream: StreamType, pdf: Any  # PdfReader
-    ) -> "NameObject":  # pragma: no cover
-        deprecate_with_replacement("readFromStream", "read_from_stream")
-        return NameObject.read_from_stream(stream, pdf)
-
-
 class DictionaryObject(dict, PdfObject):
     def raw_get(self, key: Any) -> Any:
         return dict.__getitem__(self, key)
@@ -690,7 +162,7 @@ class DictionaryObject(dict, PdfObject):
         that can be used to access XMP metadata from the document.  Can also
         return None if no metadata was found on the document root.
         """
-        from .xmp import XmpInformation
+        from ..xmp import XmpInformation
 
         metadata = self.get("/Metadata", None)
         if metadata is None:
@@ -1062,7 +534,7 @@ class StreamObject(DictionaryObject):
         stream.write(b"\nstream\n")
         data = self._data
         if encryption_key:
-            from ._security import RC4_encrypt
+            from .._security import RC4_encrypt
 
             data = RC4_encrypt(encryption_key, data)
         stream.write(data)
@@ -1094,7 +566,7 @@ class StreamObject(DictionaryObject):
         return self.flate_encode()
 
     def flate_encode(self) -> "EncodedStreamObject":
-        from .filters import FlateDecode
+        from ..filters import FlateDecode
 
         if SA.FILTER in self:
             f = self[SA.FILTER]
@@ -1144,7 +616,7 @@ class EncodedStreamObject(StreamObject):
         self.decoded_self = value
 
     def get_data(self) -> Union[None, str, bytes]:
-        from .filters import decode_stream_data
+        from ..filters import decode_stream_data
 
         if self.decoded_self is not None:
             # cached version of decoded object
@@ -1957,381 +1429,3 @@ class Bookmark(OutlineItem):  # pragma: no cover
     def __init__(self, *args: Any, **kwargs: Any) -> None:
         deprecate_with_replacement("Bookmark", "OutlineItem")
         super().__init__(*args, **kwargs)
-
-
-def createStringObject(
-    string: Union[str, bytes],
-    forced_encoding: Union[None, str, List[str], Dict[int, str]] = None,
-) -> Union[TextStringObject, ByteStringObject]:  # pragma: no cover
-    deprecate_with_replacement("createStringObject", "create_string_object", "4.0.0")
-    return create_string_object(string, forced_encoding)
-
-
-def create_string_object(
-    string: Union[str, bytes],
-    forced_encoding: Union[None, str, List[str], Dict[int, str]] = None,
-) -> Union[TextStringObject, ByteStringObject]:
-    """
-    Create a ByteStringObject or a TextStringObject from a string to represent the string.
-
-    :param Union[str, bytes] string: A string
-
-    :raises TypeError: If string is not of type str or bytes.
-    """
-    if isinstance(string, str):
-        return TextStringObject(string)
-    elif isinstance(string, bytes):
-        if isinstance(forced_encoding, (list, dict)):
-            out = ""
-            for x in string:
-                try:
-                    out += forced_encoding[x]
-                except Exception:
-                    out += bytes((x,)).decode("charmap")
-            return TextStringObject(out)
-        elif isinstance(forced_encoding, str):
-            if forced_encoding == "bytes":
-                return ByteStringObject(string)
-            return TextStringObject(string.decode(forced_encoding))
-        else:
-            try:
-                if string.startswith(codecs.BOM_UTF16_BE):
-                    retval = TextStringObject(string.decode("utf-16"))
-                    retval.autodetect_utf16 = True
-                    return retval
-                else:
-                    # This is probably a big performance hit here, but we need to
-                    # convert string objects into the text/unicode-aware version if
-                    # possible... and the only way to check if that's possible is
-                    # to try.  Some strings are strings, some are just byte arrays.
-                    retval = TextStringObject(decode_pdfdocencoding(string))
-                    retval.autodetect_pdfdocencoding = True
-                    return retval
-            except UnicodeDecodeError:
-                return ByteStringObject(string)
-    else:
-        raise TypeError("create_string_object should have str or unicode arg")
-
-
-def _create_outline_item(
-    action_ref: IndirectObject,
-    title: str,
-    color: Union[Tuple[float, float, float], str, None],
-    italic: bool,
-    bold: bool,
-) -> TreeObject:
-    outline_item = TreeObject()
-    outline_item.update(
-        {
-            NameObject("/A"): action_ref,
-            NameObject("/Title"): create_string_object(title),
-        }
-    )
-    if color:
-        if isinstance(color, str):
-            color = hex_to_rgb(color)
-        outline_item.update(
-            {NameObject("/C"): ArrayObject([FloatObject(c) for c in color])}
-        )
-    if italic or bold:
-        format_flag = 0
-        if italic:
-            format_flag += 1
-        if bold:
-            format_flag += 2
-        outline_item.update({NameObject("/F"): NumberObject(format_flag)})
-    return outline_item
-
-
-def encode_pdfdocencoding(unicode_string: str) -> bytes:
-    retval = b""
-    for c in unicode_string:
-        try:
-            retval += b_(chr(_pdfdoc_encoding_rev[c]))
-        except KeyError:
-            raise UnicodeEncodeError(
-                "pdfdocencoding", c, -1, -1, "does not exist in translation table"
-            )
-    return retval
-
-
-def decode_pdfdocencoding(byte_array: bytes) -> str:
-    retval = ""
-    for b in byte_array:
-        c = _pdfdoc_encoding[b]
-        if c == "\u0000":
-            raise UnicodeDecodeError(
-                "pdfdocencoding",
-                bytearray(b),
-                -1,
-                -1,
-                "does not exist in translation table",
-            )
-        retval += c
-    return retval
-
-
-def hex_to_rgb(value: str) -> Tuple[float, float, float]:
-    return tuple(int(value.lstrip("#")[i : i + 2], 16) / 255.0 for i in (0, 2, 4))  # type: ignore
-
-
-class AnnotationBuilder:
-    """
-    The AnnotationBuilder creates dictionaries representing PDF annotations.
-
-    Those dictionaries can be modified before they are added to a PdfWriter
-    instance via `writer.add_annotation`.
-
-    See `adding PDF annotations <../user/adding-pdf-annotations.html>`_ for
-    it's usage combined with PdfWriter.
-    """
-
-    from .types import FitType, ZoomArgType
-
-    @staticmethod
-    def text(
-        rect: Union[RectangleObject, Tuple[float, float, float, float]],
-        text: str,
-        open: bool = False,
-        flags: int = 0,
-    ) -> DictionaryObject:
-        """
-        Add text annotation.
-
-        :param Tuple[int, int, int, int] rect:
-            or array of four integers specifying the clickable rectangular area
-            ``[xLL, yLL, xUR, yUR]``
-        :param bool open:
-        :param int flags:
-        """
-        # TABLE 8.23 Additional entries specific to a text annotation
-        text_obj = DictionaryObject(
-            {
-                NameObject("/Type"): NameObject("/Annot"),
-                NameObject("/Subtype"): NameObject("/Text"),
-                NameObject("/Rect"): RectangleObject(rect),
-                NameObject("/Contents"): TextStringObject(text),
-                NameObject("/Open"): BooleanObject(open),
-                NameObject("/Flags"): NumberObject(flags),
-            }
-        )
-        return text_obj
-
-    @staticmethod
-    def free_text(
-        text: str,
-        rect: Union[RectangleObject, Tuple[float, float, float, float]],
-        font: str = "Helvetica",
-        bold: bool = False,
-        italic: bool = False,
-        font_size: str = "14pt",
-        font_color: str = "000000",
-        border_color: str = "000000",
-        background_color: str = "ffffff",
-    ) -> DictionaryObject:
-        """
-        Add text in a rectangle to a page.
-
-        :param str text: Text to be added
-        :param RectangleObject rect: or array of four integers
-            specifying the clickable rectangular area ``[xLL, yLL, xUR, yUR]``
-        :param str font: Name of the Font, e.g. 'Helvetica'
-        :param bool bold: Print the text in bold
-        :param bool italic: Print the text in italic
-        :param str font_size: How big the text will be, e.g. '14pt'
-        :param str font_color: Hex-string for the color
-        :param str border_color: Hex-string for the border color
-        :param str background_color: Hex-string for the background of the annotation
-        """
-        font_str = "font: "
-        if bold is True:
-            font_str = font_str + "bold "
-        if italic is True:
-            font_str = font_str + "italic "
-        font_str = font_str + font + " " + font_size
-        font_str = font_str + ";text-align:left;color:#" + font_color
-
-        bg_color_str = ""
-        for st in hex_to_rgb(border_color):
-            bg_color_str = bg_color_str + str(st) + " "
-        bg_color_str = bg_color_str + "rg"
-
-        free_text = DictionaryObject()
-        free_text.update(
-            {
-                NameObject("/Type"): NameObject("/Annot"),
-                NameObject("/Subtype"): NameObject("/FreeText"),
-                NameObject("/Rect"): RectangleObject(rect),
-                NameObject("/Contents"): TextStringObject(text),
-                # font size color
-                NameObject("/DS"): TextStringObject(font_str),
-                # border color
-                NameObject("/DA"): TextStringObject(bg_color_str),
-                # background color
-                NameObject("/C"): ArrayObject(
-                    [FloatObject(n) for n in hex_to_rgb(background_color)]
-                ),
-            }
-        )
-        return free_text
-
-    @staticmethod
-    def line(
-        p1: Tuple[float, float],
-        p2: Tuple[float, float],
-        rect: Union[RectangleObject, Tuple[float, float, float, float]],
-        text: str = "",
-        title_bar: str = "",
-    ) -> DictionaryObject:
-        """
-        Draw a line on the PDF.
-
-        :param Tuple[float, float] p1: First point
-        :param Tuple[float, float] p2: Second point
-        :param RectangleObject rect: or array of four
-                integers specifying the clickable rectangular area
-                ``[xLL, yLL, xUR, yUR]``
-        :param str text: Text to be displayed as the line annotation
-        :param str title_bar: Text to be displayed in the title bar of the
-            annotation; by convention this is the name of the author
-        """
-        line_obj = DictionaryObject(
-            {
-                NameObject("/Type"): NameObject("/Annot"),
-                NameObject("/Subtype"): NameObject("/Line"),
-                NameObject("/Rect"): RectangleObject(rect),
-                NameObject("/T"): TextStringObject(title_bar),
-                NameObject("/L"): ArrayObject(
-                    [
-                        FloatObject(p1[0]),
-                        FloatObject(p1[1]),
-                        FloatObject(p2[0]),
-                        FloatObject(p2[1]),
-                    ]
-                ),
-                NameObject("/LE"): ArrayObject(
-                    [
-                        NameObject(None),
-                        NameObject(None),
-                    ]
-                ),
-                NameObject("/IC"): ArrayObject(
-                    [
-                        FloatObject(0.5),
-                        FloatObject(0.5),
-                        FloatObject(0.5),
-                    ]
-                ),
-                NameObject("/Contents"): TextStringObject(text),
-            }
-        )
-        return line_obj
-
-    @staticmethod
-    def link(
-        rect: Union[RectangleObject, Tuple[float, float, float, float]],
-        border: Optional[ArrayObject] = None,
-        url: Optional[str] = None,
-        target_page_index: Optional[int] = None,
-        fit: FitType = "/Fit",
-        fit_args: Tuple[ZoomArgType, ...] = tuple(),
-    ) -> DictionaryObject:
-        """
-        Add a link to the document.
-
-        The link can either be an external link or an internal link.
-
-        An external link requires the URL parameter.
-        An internal link requires the target_page_index, fit, and fit args.
-
-
-        :param RectangleObject rect: or array of four
-            integers specifying the clickable rectangular area
-            ``[xLL, yLL, xUR, yUR]``
-        :param border: if provided, an array describing border-drawing
-            properties. See the PDF spec for details. No border will be
-            drawn if this argument is omitted.
-            - horizontal corner radius,
-            - vertical corner radius, and
-            - border width
-            - Optionally: Dash
-        :param str url: Link to a website (if you want to make an external link)
-        :param int target_page_index: index of the page to which the link should go
-                                (if you want to make an internal link)
-        :param str fit: Page fit or 'zoom' option (see below). Additional arguments may need
-            to be supplied. Passing ``None`` will be read as a null value for that coordinate.
-        :param Tuple[int, ...] fit_args: Parameters for the fit argument.
-
-
-        .. list-table:: Valid ``fit`` arguments (see Table 8.2 of the PDF 1.7 reference for details)
-           :widths: 50 200
-
-           * - /Fit
-             - No additional arguments
-           * - /XYZ
-             - [left] [top] [zoomFactor]
-           * - /FitH
-             - [top]
-           * - /FitV
-             - [left]
-           * - /FitR
-             - [left] [bottom] [right] [top]
-           * - /FitB
-             - No additional arguments
-           * - /FitBH
-             - [top]
-           * - /FitBV
-             - [left]
-        """
-        from .types import BorderArrayType
-
-        is_external = url is not None
-        is_internal = target_page_index is not None
-        if not is_external and not is_internal:
-            raise ValueError(
-                "Either 'url' or 'target_page_index' have to be provided. Both were None."
-            )
-        if is_external and is_internal:
-            raise ValueError(
-                f"Either 'url' or 'target_page_index' have to be provided. url={url}, target_page_index={target_page_index}"
-            )
-
-        border_arr: BorderArrayType
-        if border is not None:
-            border_arr = [NameObject(n) for n in border[:3]]
-            if len(border) == 4:
-                dash_pattern = ArrayObject([NameObject(n) for n in border[3]])
-                border_arr.append(dash_pattern)
-        else:
-            border_arr = [NumberObject(0)] * 3
-
-        link_obj = DictionaryObject(
-            {
-                NameObject("/Type"): NameObject("/Annot"),
-                NameObject("/Subtype"): NameObject("/Link"),
-                NameObject("/Rect"): RectangleObject(rect),
-                NameObject("/Border"): ArrayObject(border_arr),
-            }
-        )
-        if is_external:
-            link_obj[NameObject("/A")] = DictionaryObject(
-                {
-                    NameObject("/S"): NameObject("/URI"),
-                    NameObject("/Type"): NameObject("/Action"),
-                    NameObject("/URI"): TextStringObject(url),
-                }
-            )
-        if is_internal:
-            fit_arg_ready = [
-                NullObject() if a is None else NumberObject(a) for a in fit_args
-            ]
-            # This needs to be updated later!
-            dest_deferred = DictionaryObject(
-                {
-                    "target_page_index": NumberObject(target_page_index),
-                    "fit": NameObject(fit),
-                    "fit_args": ArrayObject(fit_arg_ready),
-                }
-            )
-            link_obj[NameObject("/Dest")] = dest_deferred
-        return link_obj
