+++ PyPDF2/types.py
@@ -14,14 +14,8 @@ try:
 except ImportError:
     from typing_extensions import TypeAlias
 
-from .generic import (
-    ArrayObject,
-    Destination,
-    NameObject,
-    NullObject,
-    NumberObject,
-    OutlineItem,
-)
+from .generic._base import NameObject, NullObject, NumberObject
+from .generic._data_structures import ArrayObject, Destination, OutlineItem
 
 BorderArrayType: TypeAlias = List[Union[NameObject, NumberObject, ArrayObject]]
 OutlineItemType: TypeAlias = Union[OutlineItem, Destination]
